fix additional directory support for url scenario
