+++ tools/tsp-client/src/index.ts
@@ -7,8 +7,8 @@ import { mkdir, writeFile, cp, readFile } from "node:fs/promises";
 import { addSpecFiles, checkoutCommit, cloneRepo, getRepoRoot, sparseCheckout } from "./git.js";
 import { fetch } from "./network.js";
 import { parse as parseYaml } from "yaml";
-import { joinPaths, normalizeSlashes, resolvePath } from "@typespec/compiler";
-import { formatAdditionalDirectories } from "./utils.js";
+import { joinPaths, resolvePath } from "@typespec/compiler";
+import { formatAdditionalDirectories, getAdditionalDirectoryName } from "./utils.js";
 
 
 async function sdkInit(
@@ -117,14 +117,7 @@ async function syncTspFiles(outputDir: string, localSpecRepo?: string) {
     }
     for (const dir of tspLocation.additionalDirectories!) {
       Logger.info(`Syncing additional directory: ${dir}`);
-      let normalizedDir = normalizeSlashes(dir);
-      if (normalizedDir.slice(-1) === "/") {
-        normalizedDir = normalizedDir.slice(0, -1);
-      }
-      const finalDirName = normalizedDir.split("/").pop();
-      if (!finalDirName) {
-        throw new Error(`Could not find a final directory for the following value: ${normalizedDir}`);
-      }
+      const finalDirName = getAdditionalDirectoryName(dir);
       await cp(joinPaths(localSpecRepoRoot, dir), joinPaths(tempRoot, finalDirName), { recursive: true, filter: filter });
     }
   } else {
@@ -142,10 +135,9 @@ async function syncTspFiles(outputDir: string, localSpecRepo?: string) {
     await checkoutCommit(cloneDir, tspLocation.commit);
     await cp(joinPaths(cloneDir, tspLocation.directory), srcDir, { recursive: true });
     for (const dir of tspLocation.additionalDirectories!) {
-      const dirSplit = dir.split("/");
-      let projectName = dirSplit[dirSplit.length - 1];
-      const dirName = joinPaths(tempRoot, projectName!);
-      await cp(joinPaths(cloneDir, dir), dirName, { recursive: true });
+      Logger.info(`Syncing additional directory: ${dir}`);
+      const finalDirName = getAdditionalDirectoryName(dir);
+      await cp(joinPaths(cloneDir, dir), joinPaths(tempRoot, finalDirName), { recursive: true });
     }
     Logger.debug(`Removing sparse-checkout directory ${cloneDir}`);
     await removeDirectory(cloneDir);
